Move test_product_cart to new test file
