+++ Dshop/apps/products_catalogue/tests/test_products.py
@@ -1,10 +1,8 @@
 import pytest
-from pytest import mark
 
 from lxml import etree
 
 from django.urls import reverse
-from dj_shop_cart.cart import get_cart_class
 
 from apps.products_catalogue.models import Category, Product
 
@@ -72,36 +70,3 @@ def test_generate_xml_file_for_ceneo(client):
 
     root = etree.fromstring(response.content)
     assert root.tag == 'offers'
-
-
-@mark.dj_shop_cart
-@pytest.mark.django_db
-def test_products_cart(client):
-    Cart = get_cart_class()
-
-    category = Category.objects.create(
-        name='Test Category',
-        is_active=True
-    )
-    product = Product.objects.create(
-        name='TV AMOLED',
-        price=3999.00,
-        full_description='Description 1',
-        category=category
-    )
-    quantity = 2
-    url = reverse(
-        "product-detail",
-        kwargs={
-            'slug': 'first-one',
-            'id': product.id
-        }
-    )
-    fake_response = client.get(url)
-    fake_request = fake_response.wsgi_request
-
-    cart = Cart.new(fake_request)
-    cart.add(product, quantity=quantity)
-
-    assert len(cart) == 1
-    assert cart.count == 2
